Add constants file, docstrings and comments.
Added file for constants. Maybe this should not be a thing as most of
the constants are used only in a single function call. Could maybe make
the logic into separate modules and add them there. Nevertheless, this
cleaned up main.py quite a bit.

Added comments on suggest_hero and suggest_teams. Should be a lot easier
to understand now!